+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ Change Log
 * Added a ``make`` target named ``clear-poetry-cache`` to help with certain ``poetry`` solver problems that result
   from improper caching.
 
+1.5.0
+=====
+
+* Better error diagnostics for S3FS-mounted files that are glaciated
+  if the ``CGAP_S3FS_UPLOAD_BUCKETS`` and ``CGAP_S3FS_UPLOAD_DIR`` environment variables are set.
+
+
 1.4.0
 =====